gemini/requirements.txt
google-cloud-aiplatform==1.45.0
google-auth==2.27.0
gradio==4.22.0
requests==2.31.0
vertexai==1.43.0
vobject==0.9.7
twilio==9.0.3